_stbt: Refactor: Move gstreamer utilities from core.py to gst_utils.py
...because I want to use `numpy_from_sample` from logging.py in a future
commit.
